// lib/i18n-types.ts
export type WhySectionKeys =
  | 'checklist_item1'
  | 'checklist_item2'
  | 'checklist_item3'
  | 'checklist_item4'
  | 'checklist_item5'
  | 'checklist_item6'
  | 'use_case_item1'
  | 'use_case_item2'
  | 'use_case_item3'
  | 'use_case_item4';
// Generated types for next-intl
export type WhySectionKeys = 
  | 'snippet_title'
  | 'snippet_p'
  | 'domain_list_title'
  | 'domain_list_p'
  | 'updates_title'
  | 'updates_p'
  | 'updates_item1_title'
  | 'updates_item1_desc'
  | 'updates_item2_title'
  | 'updates_item2_desc'
  | 'updates_item3_title'
  | 'updates_item3_desc'
  | 'definition_title'
  | 'definition_p'
  | 'use_case_title'
  | 'use_case_item1'
  | 'use_case_item2'
  | 'use_case_item3'
  | 'use_case_item4'
  | 'checklist_title'
  | 'checklist_item1'
  | 'checklist_item2'
  | 'checklist_item3'
  | 'checklist_item4'
  | 'checklist_item5'
  | 'checklist_item6'
  | 'checklist_p'
  | 'usage_title'
  | 'usage_p1'
  | 'usage_p2'
  | 'conclusion_title'
  | 'conclusion_p1'
  | 'conclusion_p2';

export type AppKeys = WhySectionKeys | string;